Moved functions to utlis.py. Updated readme.md.
